server/requirements.txt
bcrypt==3.2.0
certifi==2020.12.5
cffi==1.14.5
click==7.1.2
cloudinary==1.25.0
Flask==1.1.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.0.2
Flask-SQLAlchemy==2.4.4
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==2.0.1
python-dotenv==0.17.0
six==1.15.0
SQLAlchemy==1.3.23
urllib3==1.26.4
Werkzeug==1.0.1
